extract config out of adapter
